requirements.txt
Django==1.9.5
djangorestframework==3.3.3
factory-boy==2.7.0
python-dateutil==2.5.3
